examples: rename poly → stream_inband (in-band stream control pattern)
Pattern-name the directory after its control pattern (in-band control + TLAST
packetization on the AXI4-Stream); the polynomial computation keeps its identity
in the files/symbols inside (poly.py, PolyAccelComponent, build_poly_dag, the
"poly" kernel top, and the pysilicon_poly_proj Vitis project are unchanged).

git mv examples/poly → examples/stream_inband and docs/examples/poly →
docs/examples/stream_inband (6 doc pages). Per the agreed complete-scope sweep,
update every live reference to the example-directory path: build/test imports
(examples.poly.* → examples.stream_inband.*), Path("examples")/"poly"
constructions in the test suite, the root .gitignore artifact patterns, the
~20 docs/guide + docs/overview links, the moved doc pages' front-matter (title →
"Stream In-Band Control (polynomial)", teaching-order nav_order, child parent:),
and cosmetic code comments. Vitis project-dir names (pysilicon_poly_proj),
computation file/symbol names, and mcp/corpus/ are intentionally left unchanged.

The pre-existing test_dataschema_poly poly.hpp failure persists for the same
reason (poly.hpp is a generated, gitignored file), now relocated to
examples/stream_inband/poly.hpp.
