+++ examples/stream_inband/poly_evaluate_impl.tpp
