+++ examples/stream_inband/timing_analysis.py
