+++ tests/examples/test_poly_codegen.py
@@ -6,14 +6,14 @@ from pathlib import Path
 
 def test_poly_cpp_kernel_name_is_poly():
     """PolyAccelComponent overrides cpp_kernel_name to 'poly' (not 'poly_accel')."""
-    from examples.poly.poly import PolyAccelComponent
+    from examples.stream_inband.poly import PolyAccelComponent
     from pysilicon.build.hwgen import cpp_kernel_name
     assert cpp_kernel_name(PolyAccelComponent) == "poly"
 
 
 def test_poly_codegen_step_extracts_and_writes(tmp_path: Path):
     """gen_kernel writes .hpp/.cpp into <root>/gen/ and the sticky .tpp at <root>/."""
-    from examples.poly.poly_build import build_poly_dag
+    from examples.stream_inband.poly_build import build_poly_dag
     from pysilicon.build.build import BuildConfig
 
     dag = build_poly_dag()
@@ -29,7 +29,7 @@ def test_poly_codegen_step_extracts_and_writes(tmp_path: Path):
 
 def test_poly_kernel_signature_has_raw_coeffs_array():
     """CoeffArray uses cpp_storage='raw': signature has 'float coeffs[4]', not 'CoeffArray& coeffs'."""
-    from examples.poly.poly import PolyAccelComponent
+    from examples.stream_inband.poly import PolyAccelComponent
     from pysilicon.build.hwgen import kernel_signature
     from pysilicon.simulation.simulation import Simulation
 
@@ -41,7 +41,7 @@ def test_poly_kernel_signature_has_raw_coeffs_array():
 
 def test_poly_codegen_step_kernel_contains_raw_coeffs(tmp_path: Path):
     """End-to-end: generated poly.hpp contains 'coeffs[4]' in the kernel signature."""
-    from examples.poly.poly_build import build_poly_dag
+    from examples.stream_inband.poly_build import build_poly_dag
     from pysilicon.build.build import BuildConfig
 
     dag = build_poly_dag()
@@ -65,7 +65,7 @@ def test_poly_swap_over_state():
     """
     from pathlib import Path as _P
 
-    poly_root = _P(__file__).resolve().parents[2] / "examples" / "poly"
+    poly_root = _P(__file__).resolve().parents[2] / "examples" / "stream_inband"
 
     # Hand-written kernel + testbench files are gone (now codegen'd).
     assert not (poly_root / "poly.cpp").exists()
@@ -90,7 +90,7 @@ def test_poly_swap_over_state():
 
 def test_poly_gen_hpp_uses_relative_impl_include(tmp_path: Path):
     """gen/poly.hpp must reference the .tpp via ``../poly_evaluate_impl.tpp``."""
-    from examples.poly.poly_build import build_poly_dag
+    from examples.stream_inband.poly_build import build_poly_dag
     from pysilicon.build.build import BuildConfig
 
     dag = build_poly_dag()
@@ -110,7 +110,7 @@ def test_poly_gen_hpp_includes_hook_schemas_and_utility_headers(tmp_path: Path):
       (regmap field) is a ``DataArray[Float32]``, whose
       ``get_utility_includes`` returns the array-utils header path.
     """
-    from examples.poly.poly_build import build_poly_dag
+    from examples.stream_inband.poly_build import build_poly_dag
     from pysilicon.build.build import BuildConfig
 
     dag = build_poly_dag()
