+++ tests/hw/test_phase3.py
@@ -14,7 +14,7 @@ import numpy as np
 import numpy.testing as npt
 import pytest
 
-POLY_DIR = Path(__file__).resolve().parents[2] / "examples" / "poly"
+POLY_DIR = Path(__file__).resolve().parents[2] / "examples" / "stream_inband"
 if str(POLY_DIR) not in sys.path:
     sys.path.insert(0, str(POLY_DIR))
