+++ tests/poly/test_timing_analysis_plot.py
@@ -15,8 +15,8 @@ matplotlib.use("Agg")  # non-interactive backend — must be set before importin
 import matplotlib.pyplot as plt
 import pytest
 
-# ---- path setup for examples/poly sibling module ----
-_POLY_DIR = Path(__file__).resolve().parents[2] / "examples" / "poly"
+# ---- path setup for examples/stream_inband sibling module ----
+_POLY_DIR = Path(__file__).resolve().parents[2] / "examples" / "stream_inband"
 if str(_POLY_DIR) not in sys.path:
     sys.path.insert(0, str(_POLY_DIR))
